test: add unit tests for usePostZielnormRelease
RISDEV-8247
